requirements.txt
notion-client
requests
python-dotenv
pyyaml
Pillow
beautifulsoup4
tqdm
hugo
